Change sidebar to use <mat-sidenav>
Seemed wise to make use of the "angular materials" library. This lib
is already used already for the calendar widget and they also provide
a collapsible sidebar.

The use of the "flex" layouts was used to make the display full screen.
The flex displays may allow the application to move away from bootstrap
as a library. Time will tell.

Colors of the main page are messed up at this point, but the layout
looks good.