Update table in the "list guests" page to use angular material table
Moved over the self generated bootstrap table over to the angular
material table. This is a bit more consistent with the new theme we are
trying to set. The table also has may features that we would have to
build out by hand if we continued to generate the table ourselves.