Create docker file for building ui container
Creates docker file under the ci (continuous integration) directory for
building a container to host the ui. The Dockerfile makes use of a
multi-stage build using a node container as a build container. After the
build the ui is packaged into an nginx container.

Locations for these files may move around in the future as the
makerspace project doesn't have a pattern for ci in place yet.